split modules into multiple files/dirs
