+++ src/cudnn/conv/filter/filter.rs
@@ -1,59 +1,12 @@
 use core::ffi::c_void;
-use core::marker::PhantomData;
-use core::mem::MaybeUninit;
 
 use alloc::rc::Rc;
 
-use super::super::sys::*;
+use super::super::super::sys::*;
+use super::descriptor::FilterDescriptor;
 use crate::cudarc::CudaUniquePtr;
 use crate::prelude::*;
 
-pub struct FilterDescriptor<
-    T,
-    const C_OUT: usize,
-    const C_IN: usize,
-    const H: usize,
-    const W: usize,
-> {
-    descriptor: cudnnFilterDescriptor_t,
-    data_type:  PhantomData<T>,
-}
-impl<T: TensorDataType, const C_OUT: usize, const C_IN: usize, const H: usize, const W: usize>
-    FilterDescriptor<T, C_OUT, C_IN, H, W>
-{
-    pub fn create() -> CudaCudnnResult<Self> {
-        let descriptor = unsafe {
-            let mut descriptor = MaybeUninit::uninit();
-            cudnnCreateFilterDescriptor(descriptor.as_mut_ptr()).result()?;
-            descriptor.assume_init()
-        };
-        unsafe {
-            cudnnSetFilter4dDescriptor(
-                descriptor,
-                T::get_data_type(),
-                T::get_tensor_format(),
-                C_OUT as _,
-                C_IN as _,
-                H as _,
-                W as _,
-            )
-            .result()?;
-        }
-        Ok(Self {
-            descriptor,
-            data_type: PhantomData,
-        })
-    }
-}
-impl<T, const C_OUT: usize, const C_IN: usize, const H: usize, const W: usize> Drop
-    for FilterDescriptor<T, C_OUT, C_IN, H, W>
-{
-    fn drop(&mut self) {
-        unsafe { cudnnDestroyFilterDescriptor(self.descriptor) }
-            .result()
-            .unwrap();
-    }
-}
 pub struct Filter<T, const C_OUT: usize, const C_IN: usize, const H: usize, const W: usize> {
     descriptor: Rc<FilterDescriptor<T, C_OUT, C_IN, H, W>>,
     data: CudaRc<[[[[T; W]; H]; C_IN]; C_OUT]>,
@@ -75,7 +28,7 @@ where
 {
     #[inline(always)]
     pub fn get_descriptor(&self) -> cudnnFilterDescriptor_t {
-        self.descriptor.descriptor
+        self.descriptor.get_descriptor()
     }
 
     #[inline(always)]
@@ -112,27 +65,5 @@ where
         value: [[[[T; W]; H]; C_IN]; C_OUT],
     ) -> CudaCudnnResult<Self> {
         Self::create(device.take(Rc::new(value))?)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use alloc::rc::Rc;
-
-    use crate::prelude::*;
-
-    #[test]
-    fn test_create_filter() {
-        let data = [[[[0.0, 1.0]]], [[[2.0, 3.0]]]];
-        let f = Filter::create(
-            CudaDeviceBuilder::new(0)
-                .build()
-                .unwrap()
-                .take(Rc::new(data))
-                .unwrap(),
-        )
-        .unwrap();
-        let on_gpu = *f.data.sync_release().unwrap().unwrap();
-        assert_eq!(data, on_gpu);
     }
 }
